refactor: move lint-controller and lint-fixes into lint/ directory
- Move src/wiki/lint-controller.ts to src/wiki/lint/controller.ts
- Move src/wiki/lint-fixes.ts to src/wiki/lint/fixer.ts
- Update all import paths in: main.ts, wiki-engine.ts, page-factory.ts,
  source-analyzer.ts, lint/controller.ts, lint/fixer.ts, lint/phases/*.ts.
- Update test import paths: lint-fixes.test.ts, lint/fix-runners.test.ts.
- Add LLMWikiSettings to fixer.ts imports (was previously dynamic
  import({../types}) which tripped @typescript-eslint/no-unsafe-argument).
- 4-Gate clean; 728 tests passing.
- No public API changes; fix-runners.ts already worked independent of
  LintFixer class, so the new fixer.ts only serves wiki-engine.ts.
